Allow more types of nightly jobs

+++ Scripts/nightly_jobs.sh
@@ -8,11 +8,14 @@
 usage() { echo "Usage: $0
   --dir : Directory under Production/Validation/nightly to find the script
   --script: Script name in the above directory to process
-  --dataset : Dataset to process
+  [ --type "jobtype"]: either primary, resample, or read (default)
+  [ --dataset : Dataset to process (if needed) ]
+  [ --resample : Resampler stream ]
   [ --location "location" ] : location to fine dataset (default tape)
   [ --submit / --nosubmit] : submit the jobs or just create the fcl (default nosubmit)
   [ --jobs N ] : Number of jobs to run (default 10)
   [ --merge N ] : Merge N inputs to 1 output (default 1)
+  [ --run N --events M ] : Specify run # and events/job when no input dataset
   [ --help ] : Print this message.
   Note: Muse must be setup to point to the nightly build for this script to run"
 }
@@ -26,12 +29,17 @@ NJOBS=10
 MERGE=1
 SUBMIT=""
 SCRIPT=""
+TYPE="read"
 DATASET=""
+RESAMPLE=""
 DATE=`date --iso-8601`
 LOC="tape"
+RUN=1202
+EVENTS=1000
 NIGHTLY="/exp/mu2e/app/users/mu2epro/nightly2/current"
 NUMBERS='^[0-9]+$'
 declare -a LOCATIONS=("tape" "disk")
+declare -a TYPES=("primary" "resample" "read")
 
 while getopts ":-:h" LONGOPT; do
   case "${LONGOPT}" in
@@ -43,9 +51,15 @@ while getopts ":-:h" LONGOPT; do
         script)
           SCRIPT=${!OPTIND} OPTIND=$(( $OPTIND + 1 ))
           ;;
+        type)
+          TYPE=${!OPTIND} OPTIND=$(( $OPTIND + 1 ))
+          ;;
         dataset)
           DATASET=${!OPTIND} OPTIND=$(( $OPTIND + 1 ))
           # add a test that the dataset exists TODO
+          ;;
+        resample)
+          RESAMPLE=${!OPTIND} OPTIND=$(( $OPTIND + 1 ))
           ;;
         location)
           LOC=${!OPTIND} OPTIND=$(( $OPTIND + 1 ))
@@ -55,6 +69,12 @@ while getopts ":-:h" LONGOPT; do
           ;;
         nosubmit)
           SUBMIT=""
+          ;;
+        events)
+          EVENTS=${!OPTIND} OPTIND=$(( $OPTIND + 1 ))
+          ;;
+        run)
+          RUN=${!OPTIND} OPTIND=$(( $OPTIND + 1 ))
           ;;
         jobs)
           NJOBS=${!OPTIND} OPTIND=$(( $OPTIND + 1 ))
@@ -97,14 +117,23 @@ while getopts ":-:h" LONGOPT; do
       ;;
   esac
 done
-# Test inputs
 FULLSCRIPT=Production/Validation/nightly/${VDIR}/${SCRIPT}.fcl
 if [[ ! -f ${FULLSCRIPT} ]]; then
   echo "Validation script ${FULLSCRIPT} does not exist!"
   exit_abnormal
 fi
-if [[ ${DATASET} == "" ]]; then
-  echo "Dataset unset"
+goodtype="no"
+for type in ${TYPES[@]}; do
+  if [[ ${type} == ${TYPE} ]]; then
+    goodtype="yes"
+    echo "Preparing jobs of type ${TYPE}"
+  fi
+done
+if [[ ${goodtype} == "no" ]]; then
+  echo "Bad job type ${TYPE}; should be one of"
+  for type in ${TYPES[@]}; do
+    echo ${type}
+  done
   exit_abnormal
 fi
 goodloc="no"
@@ -135,24 +164,38 @@ fi
 # tools needed
 setup mu2efiletools
 setup mu2egrid
-
 OUTDIR="/pnfs/mu2e/scratch/users/${USER}/${VDIR}Validation"
 if [[ ! -d ${OUTDIR} ]]; then
   mkdir ${OUTDIR}
 fi
-
-INPUTS=${OUTDIR}/dts.${VDIR}${SCRIPT}.txt
-if [[ -f ${INPUTS} ]]; then
-  rm -f ${INPUTS}
+if [[ ${DATASET} == "" ]]; then
+# empty input, job must set run number and # of events
+  RUN=1202
+  EVENTS=1000
+else
+  INPUTS=${OUTDIR}/dts.${VDIR}${SCRIPT}.txt
+  if [[ -f ${INPUTS} ]]; then
+    rm -f ${INPUTS}
+  fi
+ mu2eDatasetFileList --basename ${DATASET} > ${INPUTS}
 fi
 
-mu2eDatasetFileList --basename ${DATASET} > ${INPUTS}
 JOBDEF="${OUTDIR}/cnf.${USER}.${VDIR}.${SCRIPT}.0.tar"
 if [[ -f ${JOBDEF} ]]; then
   rm -f ${JOBDEF}
 fi
 # create the job definition
-mu2ejobdef --code ${NIGHTLYBUILD}/${DATE}.tgz --description ${VDIR} --dsconf ${SCRIPT} --dsowner ${USER} --inputs ${INPUTS} --merge-factor ${MERGE} --embed ${FULLSCRIPT} --outdir ${OUTDIR}
+if [[ ${TYPE} == "primary" ]]; then
+    mu2ejobdef --code ${NIGHTLYBUILD}/${DATE}.tgz --description ${VDIR} --dsconf ${SCRIPT} --dsowner ${USER} --run-number=${RUN} --events-per-job=${EVENTS} --embed ${FULLSCRIPT} --outdir ${OUTDIR}
+else
+  if [[ ${TYPE} == "resample" ]]; then
+    AUXINPUT=${MERGE}:physics.filters.${RESAMPLE}.fileNames:${INPUTS}
+    echo "Resampling ${AUXINPUT}"
+    mu2ejobdef --code ${NIGHTLYBUILD}/${DATE}.tgz --description ${VDIR} --dsconf ${SCRIPT} --dsowner ${USER} --auxinput ${AUXINPUT} --run-number=${RUN} --events-per-job=${EVENTS} --embed ${FULLSCRIPT} --outdir ${OUTDIR}
+  else
+    mu2ejobdef --code ${NIGHTLYBUILD}/${DATE}.tgz --description ${VDIR} --dsconf ${SCRIPT} --dsowner ${USER} --inputs ${INPUTS} --merge-factor ${MERGE} --embed ${FULLSCRIPT} --outdir ${OUTDIR}
+  fi
+fi
 if [[ ${SUBMIT} == "yes" ]];  then
   echo "Submitting ${NJOBS} jobs to the grid using ${JOBDEF}"
   mu2ejobsub --jobdef ${JOBDEF} --default-protocol ifdh --default-location ${LOC} --firstjob 0 --njobs ${NJOBS} --role production --memory 2GB
